+++ src/mcp/index.ts
@@ -3,30 +3,48 @@ import { StdioServerTransport } from "@modelcontextprotocol/sdk/server/stdio.js"
 import {
   CallToolRequest,
   CallToolRequestSchema,
-  CallToolResult,
-  ListToolsRequestSchema,
   ListToolsResult,
+  LoggingLevel,
+  SetLevelRequestSchema,
+  ListToolsRequestSchema,
+  CallToolResult,
 } from "@modelcontextprotocol/sdk/types.js";
-import { checkFeatureActive, mcpError } from "./util.js";
-import { ClientConfig, SERVER_FEATURES, ServerFeature } from "./types.js";
-import { availableTools } from "./tools/index.js";
-import { ServerTool, ServerToolContext } from "./tool.js";
-import { configstore } from "../configstore.js";
-import { Command } from "../command.js";
-import { requireAuth } from "../requireAuth.js";
-import { Options } from "../options.js";
-import { getProjectId } from "../projectUtils.js";
-import { mcpAuthError, NO_PROJECT_ERROR } from "./errors.js";
-import { trackGA4 } from "../track.js";
-import { Config } from "../config.js";
-import { loadRC } from "../rc.js";
-import { EmulatorHubClient } from "../emulator/hubClient.js";
-import { Emulators } from "../emulator/types.js";
+import { checkFeatureActive, mcpError } from "./util";
+import { ClientConfig, SERVER_FEATURES, ServerFeature } from "./types";
+import { availableTools } from "./tools/index";
+import { ServerTool, ServerToolContext } from "./tool";
+import { configstore } from "../configstore";
+import { Command } from "../command";
+import { requireAuth } from "../requireAuth";
+import { Options } from "../options";
+import { getProjectId } from "../projectUtils";
+import { mcpAuthError, NO_PROJECT_ERROR, mcpGeminiError } from "./errors";
+import { trackGA4 } from "../track";
+import { Config } from "../config";
+import { loadRC } from "../rc";
+import { EmulatorHubClient } from "../emulator/hubClient";
+import { Emulators } from "../emulator/types";
 import { existsSync } from "node:fs";
+import { ensure, check } from "../ensureApiEnabled";
+import * as api from "../api";
+import { LoggingStdioServerTransport } from "./logging-transport";
+import { isFirebaseStudio } from "../env";
+import { timeoutFallback } from "../timeout";
 
-const SERVER_VERSION = "0.1.0";
+const SERVER_VERSION = "0.2.0";
 
-const cmd = new Command("experimental:mcp").before(requireAuth);
+const cmd = new Command("experimental:mcp");
+
+const orderedLogLevels = [
+  "debug",
+  "info",
+  "notice",
+  "warning",
+  "error",
+  "critical",
+  "alert",
+  "emergency",
+] as const;
 
 export class FirebaseMcpServer {
   private _ready: boolean = false;
@@ -39,21 +57,43 @@ export class FirebaseMcpServer {
   clientInfo?: { name?: string; version?: string };
   emulatorHubClient?: EmulatorHubClient;
 
+  // logging spec:
+  // https://modelcontextprotocol.io/specification/2025-03-26/server/utilities/logging
+  currentLogLevel?: LoggingLevel = process.env.FIREBASE_MCP_DEBUG_LOG ? "debug" : undefined;
+  // the api of logging from a consumers perspective looks like `server.logger.warn("my warning")`.
+  public readonly logger = Object.fromEntries(
+    orderedLogLevels.map((logLevel) => [
+      logLevel,
+      (message: unknown) => this.log(logLevel, message),
+    ]),
+  ) as Record<LoggingLevel, (message: unknown) => Promise<void>>;
+
+  /** Create a special tracking function to avoid blocking everything on initialization notification. */
+  private async trackGA4(
+    event: Parameters<typeof trackGA4>[0],
+    params: Parameters<typeof trackGA4>[1] = {},
+  ): Promise<void> {
+    // wait until ready or until 2s has elapsed
+    if (!this.clientInfo) await timeoutFallback(this.ready(), null, 2000);
+    const clientInfoParams = {
+      mcp_client_name: this.clientInfo?.name || "<unknown-client>",
+      mcp_client_version: this.clientInfo?.version || "<unknown-version>",
+    };
+    trackGA4(event, { ...params, ...clientInfoParams });
+  }
+
   constructor(options: { activeFeatures?: ServerFeature[]; projectRoot?: string }) {
     this.activeFeatures = options.activeFeatures;
     this.startupRoot = options.projectRoot || process.env.PROJECT_ROOT;
     this.server = new Server({ name: "firebase", version: SERVER_VERSION });
-    this.server.registerCapabilities({ tools: { listChanged: true } });
+    this.server.registerCapabilities({ tools: { listChanged: true }, logging: {} });
     this.server.setRequestHandler(ListToolsRequestSchema, this.mcpListTools.bind(this));
     this.server.setRequestHandler(CallToolRequestSchema, this.mcpCallTool.bind(this));
     this.server.oninitialized = async () => {
       const clientInfo = this.server.getClientVersion();
       this.clientInfo = clientInfo;
       if (clientInfo?.name) {
-        trackGA4("mcp_client_connected", {
-          mcp_client_name: clientInfo.name,
-          mcp_client_version: clientInfo.version,
-        });
+        this.trackGA4("mcp_client_connected");
       }
       if (!this.clientInfo?.name) this.clientInfo = { name: "<unknown-client>" };
 
@@ -62,6 +102,12 @@ export class FirebaseMcpServer {
         this._readyPromises.pop()?.resolve();
       }
     };
+
+    this.server.setRequestHandler(SetLevelRequestSchema, async ({ params }) => {
+      this.currentLogLevel = params.level;
+      return {};
+    });
+
     this.detectProjectRoot();
     this.detectActiveFeatures();
   }
@@ -74,8 +120,12 @@ export class FirebaseMcpServer {
     });
   }
 
+  get clientName(): string {
+    return this.clientInfo?.name ?? (isFirebaseStudio() ? "Firebase Studio" : "<unknown-client>");
+  }
+
   private get clientConfigKey() {
-    return `mcp.clientConfigs.${this.clientInfo?.name || "<unknown-client>"}:${this.startupRoot || process.cwd()}`;
+    return `mcp.clientConfigs.${this.clientName}:${this.startupRoot || process.cwd()}`;
   }
 
   getStoredClientConfig(): ClientConfig {
@@ -90,15 +140,17 @@ export class FirebaseMcpServer {
   }
 
   async detectProjectRoot(): Promise<string> {
-    await this.ready();
+    await timeoutFallback(this.ready(), null, 2000);
     if (this.cachedProjectRoot) return this.cachedProjectRoot;
     const storedRoot = this.getStoredClientConfig().projectRoot;
     this.cachedProjectRoot = storedRoot || this.startupRoot || process.cwd();
+    this.log("debug", "detected and cached project root: " + this.cachedProjectRoot);
     return this.cachedProjectRoot;
   }
 
   async detectActiveFeatures(): Promise<ServerFeature[]> {
     if (this.detectedFeatures?.length) return this.detectedFeatures; // memoized
+    this.log("debug", "detecting active features of Firebase MCP server...");
     const options = await this.resolveOptions();
     const projectId = await this.getProjectId();
     const detected = await Promise.all(
@@ -108,6 +160,10 @@ export class FirebaseMcpServer {
       }),
     );
     this.detectedFeatures = detected.filter((f) => !!f) as ServerFeature[];
+    this.log(
+      "debug",
+      "detected features of Firebase MCP server: " + (this.detectedFeatures.join(", ") || "<none>"),
+    );
     return this.detectedFeatures;
   }
 
@@ -172,11 +228,14 @@ export class FirebaseMcpServer {
     return getProjectId(await this.resolveOptions());
   }
 
-  async getAuthenticatedUser(): Promise<string | null> {
+  async getAuthenticatedUser(skipAutoAuth: boolean = false): Promise<string | null> {
     try {
-      const email = await requireAuth(await this.resolveOptions());
-      return email ?? "Application Default Credentials";
+      this.log("debug", `calling requireAuth`);
+      const email = await requireAuth(await this.resolveOptions(), skipAutoAuth);
+      this.log("debug", `detected authenticated account: ${email || "<none>"}`);
+      return email ?? (skipAutoAuth ? null : "Application Default Credentials");
     } catch (e) {
+      this.log("debug", `error in requireAuth: ${e}`);
       return null;
     }
   }
@@ -184,16 +243,15 @@ export class FirebaseMcpServer {
   async mcpListTools(): Promise<ListToolsResult> {
     await Promise.all([this.detectActiveFeatures(), this.detectProjectRoot()]);
     const hasActiveProject = !!(await this.getProjectId());
-    await trackGA4("mcp_list_tools", {
-      mcp_client_name: this.clientInfo?.name,
-      mcp_client_version: this.clientInfo?.version,
-    });
+    await this.trackGA4("mcp_list_tools");
+    const skipAutoAuthForStudio = isFirebaseStudio();
+    this.log("debug", `skip auto-auth in studio environment: ${skipAutoAuthForStudio}`);
     return {
       tools: this.availableTools.map((t) => t.mcp),
       _meta: {
         projectRoot: this.cachedProjectRoot,
         projectDetected: hasActiveProject,
-        authenticatedUser: await this.getAuthenticatedUser(),
+        authenticatedUser: await this.getAuthenticatedUser(skipAutoAuthForStudio),
         activeFeatures: this.activeFeatures,
         detectedFeatures: this.detectedFeatures,
       },
@@ -207,6 +265,7 @@ export class FirebaseMcpServer {
     const tool = this.getTool(toolName);
     if (!tool) throw new Error(`Tool '${toolName}' could not be found.`);
 
+    // Check if the current project directory exists.
     if (
       tool.mcp.name !== "firebase_update_environment" && // allow this tool only, to fix the issue
       (!this.cachedProjectRoot || !existsSync(this.cachedProjectRoot))
@@ -215,15 +274,29 @@ export class FirebaseMcpServer {
         `The current project directory '${this.cachedProjectRoot || "<NO PROJECT DIRECTORY FOUND>"}' does not exist. Please use the 'update_firebase_environment' tool to target a different project directory.`,
       );
     }
+
+    // Check if the project ID is set.
     let projectId = await this.getProjectId();
     if (tool.mcp._meta?.requiresProject && !projectId) {
       return NO_PROJECT_ERROR;
     }
     projectId = projectId || "";
 
+    // Check if the user is logged in.
     const accountEmail = await this.getAuthenticatedUser();
     if (tool.mcp._meta?.requiresAuth && !accountEmail) {
       return mcpAuthError();
+    }
+
+    // Check if the tool requires Gemini in Firebase API.
+    if (tool.mcp._meta?.requiresGemini) {
+      if (configstore.get("gemini")) {
+        await ensure(projectId, api.cloudAiCompanionOrigin(), "");
+      } else {
+        if (!(await check(projectId, api.cloudAiCompanionOrigin(), ""))) {
+          return mcpGeminiError(projectId);
+        }
+      }
     }
 
     const options = { projectDir: this.cachedProjectRoot, cwd: this.cachedProjectRoot };
@@ -236,26 +309,44 @@ export class FirebaseMcpServer {
     };
     try {
       const res = await tool.fn(toolArgs, toolsCtx);
-      await trackGA4("mcp_tool_call", {
+      await this.trackGA4("mcp_tool_call", {
         tool_name: toolName,
         error: res.isError ? 1 : 0,
-        mcp_client_name: this.clientInfo?.name,
-        mcp_client_version: this.clientInfo?.version,
       });
       return res;
     } catch (err: unknown) {
-      await trackGA4("mcp_tool_call", {
+      await this.trackGA4("mcp_tool_call", {
         tool_name: toolName,
         error: 1,
-        mcp_client_name: this.clientInfo?.name,
-        mcp_client_version: this.clientInfo?.version,
       });
       return mcpError(err);
     }
   }
 
   async start(): Promise<void> {
-    const transport = new StdioServerTransport();
+    const transport = process.env.FIREBASE_MCP_DEBUG_LOG
+      ? new LoggingStdioServerTransport(process.env.FIREBASE_MCP_DEBUG_LOG)
+      : new StdioServerTransport();
     await this.server.connect(transport);
+  }
+
+  private async log(level: LoggingLevel, message: unknown) {
+    let data = message;
+
+    // mcp protocol only takes jsons or it errors; for convienence, format
+    // a a string into a json.
+    if (typeof message === "string") {
+      data = { message };
+    }
+
+    if (!this.currentLogLevel) {
+      return;
+    }
+
+    if (orderedLogLevels.indexOf(this.currentLogLevel) > orderedLogLevels.indexOf(level)) {
+      return;
+    }
+
+    if (this._ready) await this.server.sendLoggingMessage({ level, data });
   }
 }