+++ firebase-vscode/CHANGELOG.md
@@ -1,6 +1,23 @@
 ## NEXT
 
+## 1.6.0
+
+- Update internal `firebase-tools` dependency to 14.11.1
+- [Changed] Now integrates with GCA in its agentic mode, powered by the Gemini CLI. This brings the Firebase MCP Server directly into the VS Code environment, enabling developers to use natural language to generate application schemas and queries without manually invoking explicit tools.
+
+## 1.5.1
+
+- Update internal `firebase-tools` dependency to 14.11.0
+- [Fixed] Language server now properly recognizes nested Dataconnect folders
+- [Fixed] Add Data and Read Data now properly support enum and list types
+
+## 1.5.0
+
+- Update internal `firebase-tools` dependency to 14.9.0
+- Update internal `graphql-language-server` dependency
+- Update internal `graphql-language-service-server` dependency
 - [Changed] Graphql Syntax Highlighter is now an extension dependency
+- [Fixed] Language server now recognizes fragments in other files
 
 ## 1.4.1